+++ CHANGELOG.md
@@ -4,6 +4,23 @@
 
 ### Added
 
+* Added creature trophy icons to tracked creature map pins.
+* Added support for manually removing tracked creature map pins by right-clicking them.
+* Added a configuration option to enable or disable manual map pin removal.
+* Added tracking for Gammeltrolls.
+* Added tracking for Moose.
+* Added tracking for Fallen Warriors.
+* Added tracking for Eyeless Ones.
+* Added tracking for Elaking.
+* Added tracking for White Deer.
+* Added tracking for Seeker Soldiers.
+* Added tracking for 1-star or higher Seekers.
+* Added tracking for Gjalls.
+
+### Changed
+
+* Organized tracked creature configuration settings by biome.
+
 ## v1.2.0
 
 ### Added